Improved project structure and removed StyleCop warnings
