+++ src/BaloonsPop.Common/Validators/IMatrixValidator.cs
@@ -1,0 +1,7 @@
+﻿namespace BaloonsPop.Common.Validators
+{
+    public interface IMatrixValidator
+    {
+        bool IsInsideMatrix<T>(T[,] matrix, int row, int col);
+    }
+}
