+++ src/BaloonsPop.Common/Validators/IUserInputValidator.cs
@@ -1,0 +1,7 @@
+﻿namespace BaloonsPop.Common.Validators
+{
+    public interface IUserInputValidator
+    {
+        bool IsValidUserMove(string userInput);
+    }
+}
