+++ src/BaloonsPop.Common/Validators/UserInputValidator.cs
@@ -1,0 +1,46 @@
+﻿namespace BaloonsPop.Common.Validators
+{
+    public class UserInputValidator : IUserInputValidator
+    {
+        private const char COMMA = ',';
+        private const char DOT = '.';
+        private const int VALID_INPUT_LENGTH = 3;
+        private const int MAX_ROW_INPUT_VALUE = 4;
+
+        private static readonly UserInputValidator Instance = new UserInputValidator();
+
+        private UserInputValidator()
+        {
+        }
+
+        public static UserInputValidator GetInstance
+        {
+            get
+            {
+                return UserInputValidator.Instance;
+            }
+        }
+
+        public bool IsValidUserMove(string userInput)
+        {
+            bool hasCorrectLength = userInput.Length == VALID_INPUT_LENGTH;
+
+            if (!hasCorrectLength)
+            {
+                return false;
+            }
+
+            bool firstCharIsDigit = char.IsDigit(userInput[2]);
+            bool secondCharIsValid = char.IsWhiteSpace(userInput[1]) || userInput[1] == DOT || userInput[1] == COMMA;
+            bool thirdCharIsDigit = char.IsDigit(userInput[2]);
+
+            return firstCharIsDigit && secondCharIsValid && thirdCharIsDigit && this.IsValidRowMove(userInput[0]);
+        }
+
+        private bool IsValidRowMove(char notParsedRow)
+        {
+            bool isValidRow = (notParsedRow - 48) <= MAX_ROW_INPUT_VALUE;
+            return isValidRow;
+        }
+    }
+}
