+++ src/BaloonsPop.ConsoleUI/Properties/AssemblyInfo.cs
