+++ src/BaloonsPop.Engine/GameLogic.cs
@@ -1,0 +1,113 @@
+﻿namespace BaloonsPop.Engine
+{
+    using System;
+    using System.Collections.Generic;
+    using BaloonsPop.Common.Validators;
+
+    public class GameLogic
+    {
+        private const int FIELD_ROWS = 4;
+        private const int FIELD_COLS = 9;
+        private const int MIN_BALOON_VALUE = 1;
+        private const int MAX_BALOON_VALUE = 4;
+
+        private static readonly int[][] PopDirections = new int[][] { new int[] { 0, 1 }, new int[] { 0, -1 }, new int[] { 1, 0 }, new int[] { -1, 0 } };
+
+        private Random rng;
+
+        private IMatrixValidator matrixValidator;
+
+        public GameLogic(IMatrixValidator matrixValidator)
+        {
+            this.matrixValidator = matrixValidator;
+            this.rng = new Random();
+        }
+
+        public byte[,] GenerateField()
+        {
+            var newField = new byte[FIELD_ROWS + 1, FIELD_COLS + 1];
+
+            for (var row = 0; row <= FIELD_ROWS; row++)
+            {
+                for (var column = 0; column <= FIELD_COLS; column++)
+                {
+                    var currentBaloonValue = this.GetRandomBaloonValue();
+                    newField[row, column] = currentBaloonValue;
+                }
+            }
+
+            return newField;
+        }
+
+        public void PopBaloons(byte[,] baloonField, int row, int column)
+        {
+            foreach (var dir in PopDirections)
+            {
+                this.PopInDirection(baloonField, row, column, dir[0], dir[1]);
+            }
+
+            baloonField[row, column] = 0;
+        }
+
+        public void LetBaloonsFall(byte[,] baloonField)
+        {
+            var baloonColumn = new Stack<byte>();
+
+            for (int column = 0, length = baloonField.GetLength(1); column < length; column++)
+            {
+                for (int row = 0, rowsCount = baloonField.GetLength(0); row < rowsCount; row++)
+                {
+                    if (baloonField[row, column] != 0)
+                    {
+                        baloonColumn.Push(baloonField[row, column]);
+                    }
+                }
+
+                for (int row = baloonField.GetLength(0) - 1; row >= 0; row--)
+                {
+                    if (baloonColumn.Count > 0)
+                    {
+                        baloonField[row, column] = baloonColumn.Pop();
+                    }
+                    else
+                    {
+                        baloonField[row, column] = 0;
+                    }
+                }
+            }
+        }
+
+        public bool GameIsOver(byte[,] matrix)
+        {
+            foreach (var cell in matrix)
+            {
+                if (cell != 0)
+                {
+                    return false;
+                }
+            }
+
+            return true;
+        }
+
+        private void PopInDirection(byte[,] matrix, int row, int col, int xUpdate, int yUpdate)
+        {
+            var baloonType = matrix[row, col];
+            row += yUpdate;
+            col += xUpdate;
+
+            while (this.matrixValidator.IsInsideMatrix(matrix, row, col) && matrix[row, col] == baloonType)
+            {
+                matrix[row, col] = 0;
+                row += yUpdate;
+                col += xUpdate;
+            }
+        }
+
+        private byte GetRandomBaloonValue()
+        {
+            var randomBaloonValue = (byte)this.rng.Next(MIN_BALOON_VALUE, MAX_BALOON_VALUE + 1);
+            return randomBaloonValue;
+        }
+    }
+}
