+++ src/BaloonsPop.Engine/IEngine.cs
@@ -1,0 +1,11 @@
+﻿namespace BaloonsPop.Engine
+{
+    using BaloonsPop.Common.Validators;
+
+    public interface IEngine
+    {
+        void Run();
+
+        void Initialize(IUserInterface ui, IUserInputValidator validator);
+    }
+}
