+++ src/BaloonsPop.Engine/IPrinter.cs
@@ -1,0 +1,11 @@
+﻿namespace BaloonsPop.Engine
+{
+    public interface IPrinter
+    {
+        void PrintMessage(string message);
+
+        void PrintField(byte[,] matrix);
+
+        void PrintHighscore(string highscore);
+    }
+}
