+++ src/BaloonsPop.Engine/IUserInputReader.cs
@@ -1,0 +1,7 @@
+﻿namespace BaloonsPop.Engine
+{
+    public interface IUserInputReader
+    {
+        string ReadUserInput();
+    }
+}
