+++ src/BaloonsPop.Engine/IUserInterface.cs
@@ -1,0 +1,6 @@
+﻿namespace BaloonsPop.Engine
+{
+    public interface IUserInterface : IUserInputReader, IPrinter
+    {
+    }
+}
